requirements.txt
asgiref==3.6.0
Django==4.1.7
django-autoslug==1.9.9
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2023.3
